exhibition-screen linting
