+++ package.json
@@ -49,6 +49,7 @@
     "bundle-report": "npm run ng-high-memory -- build --configuration production --aot --stats-json --named-chunks=true && webpack-bundle-analyzer dist/browser/stats.json",
     "lint": "eslint \"projects/laji/src/**/*.ts\"",
     "lint:bird-atlas": "eslint \"projects/bird-atlas/src/**/*.ts\"",
+    "lint:exhibition-screen": "eslint \"projects/exhibition-screen/src/**/*.ts\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor projects/laji/e2e/protractor.conf.js",
