+++ projects/exhibition-screen/.eslintrc.json
@@ -1,0 +1,23 @@
+{
+  "extends": [
+    "../../.eslintrc.json"
+  ],
+  "rules": {
+    "@angular-eslint/component-selector": [
+      "error",
+      {
+        "type": "element",
+        "prefix": ["es"],
+        "style": "kebab-case"
+      }
+    ],
+    "@angular-eslint/directive-selector": [
+      "error",
+      {
+        "type": "attribute",
+        "prefix": ["es"],
+        "style": "camelCase"
+      }
+    ]
+  }
+}
