+++ projects/exhibition-screen/src/animations/animatables/es-animatable.ts
@@ -1,0 +1,7 @@
+import { Deg, V2 } from '../math';
+
+export interface EsAnimatable {
+  pos: V2;
+  rot: Deg;
+  update(dt: number): boolean;
+}
