+++ projects/exhibition-screen/src/app/app.module.ts
@@ -1,7 +1,6 @@
-import { Injectable, NgModule } from '@angular/core';
-import { BrowserModule, HammerGestureConfig, HammerModule, HAMMER_GESTURE_CONFIG } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { BrowserModule, HammerModule, HAMMER_GESTURE_CONFIG } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
-import * as Hammer from 'hammerjs';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { SlideshowComponent } from './slideshow/slideshow.component';
 import { SlideComponent } from './slideshow/slide/slide.component';
@@ -10,14 +9,7 @@ import { TranslateModule } from '@ngx-translate/core';
 import { HttpClientModule } from '@angular/common/http';
 import { SlideNavigationComponent } from './slideshow/slide-navigation/slide-navigation.component';
 import { PdfModalComponent } from './slideshow/slide/pdf-modal.component';
-
-@Injectable()
-export class HammerConfig extends HammerGestureConfig {
-  overrides = <any> {
-    swipe: { direction: Hammer.DIRECTION_ALL },
-    pan: { direction: Hammer.DIRECTION_ALL },
-  };
-}
+import { HammerConfig } from './hammer-config';
 
 @NgModule({
   imports: [
