+++ projects/exhibition-screen/src/app/core/i18n-map.ts
@@ -2,8 +2,8 @@ export type Lang = 'fi' | 'en' | 'sv';
 
 export const i18nMap: {[key: string]: {[key in Lang]: string}} = {
   screenOne: {
-    'fi': 'ex-12',
-    'en': 'ex-16',
-    'sv': 'ex-14'
+    fi: 'ex-12',
+    en: 'ex-16',
+    sv: 'ex-14'
   }
 };
