+++ projects/exhibition-screen/src/app/hammer-config.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+import { HammerGestureConfig } from '@angular/platform-browser';
+import * as Hammer from 'hammerjs';
+
+@Injectable()
+export class HammerConfig extends HammerGestureConfig {
+  overrides = <any> {
+    swipe: { direction: Hammer.DIRECTION_ALL },
+    pan: { direction: Hammer.DIRECTION_ALL },
+  };
+}
