+++ projects/exhibition-screen/src/app/slideshow/slide/pdf-modal.component.ts
@@ -1,5 +1,5 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input } from "@angular/core";
-import { DomSanitizer, SafeResourceUrl } from "@angular/platform-browser";
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input } from '@angular/core';
+import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 
 @Component({
   selector: 'es-pdf-modal',
@@ -11,7 +11,7 @@ import { DomSanitizer, SafeResourceUrl } from "@angular/platform-browser";
 export class PdfModalComponent {
   public safeUrl: SafeResourceUrl = this.sanitizer.bypassSecurityTrustResourceUrl('');
   @Input() set url(u: string) {
-    this.safeUrl = this.sanitizer.bypassSecurityTrustResourceUrl(u + "#toolbar=0&navpanes=0");
+    this.safeUrl = this.sanitizer.bypassSecurityTrustResourceUrl(u + '#toolbar=0&navpanes=0');
     this.cdr.markForCheck();
   }
