--- projects/exhibition-screen/tslint.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tslint.json",
-  "rules": {
-    "directive-selector": [
-      true,
-      "attribute",
-      "vir",
-      "camelCase"
-    ],
-    "component-selector": [
-      true,
-      "element",
-      "vir",
-      "kebab-case"
-    ]
-  }
-}
